+++ async_API.h
@@ -42,6 +42,8 @@ struct _async_await_context_t
 	bool ignore_errors;
 	/* If we need to fill missing results with null */
 	bool fill_missing_with_null;
+	/* If we need to preserve key order in results */
+	bool preserve_key_order;
 	/*
 	 * The flag indicates that all pending coroutines
 	 * must be cancelled once the wait completes, regardless of the outcome.
@@ -93,6 +95,7 @@ void async_await_futures(
 	HashTable *results,
 	HashTable *errors,
 	bool fill_missing_with_null,
+	bool preserve_key_order,
 	bool cancel_on_exit
 );